main.py
import discord
from dotenv import load_dotenv
import os




class MyClient(discord.Client):
    async def on_ready(self):
        print(f'Logged on as {self.user}!')

    async def on_message(self, message):
        print(f'Message from {message.author}: {message.content}')


def init():
    load_dotenv()

    intents = discord.Intents.default()
    intents.message_content = True

    client = MyClient(intents=intents)
    client.run(os.getenv('DISCORD_TOKEN'))


if __name__ == "__main__":
    init()